refactor assert_eq! assert_ne!
