+++ macros/src/functions.rs
@@ -1,3 +1,4 @@
+pub(crate) mod assert_binop;
 pub(crate) mod assert_like;
 pub(crate) mod dbg;
 pub(crate) mod panic_like;
